added gitignore. separated web server class from esp32 server file
